Time: 95 ms (16.93%), Space: 29.2 MB (51.24%) - LeetHub

+++ 0643-maximum-average-subarray-i/0643-maximum-average-subarray-i.py
@@ -8,7 +8,7 @@ class Solution:
             total += nums[r]
 
             if (r - l + 1) == k:
-                res = max(res, float(total / k))
+                res = max(res, total / k)
                 total -= nums[l]
                 l += 1
